Make submission date optional
Metadata typically includes the 'submission_date' field, which was used
in two rules: `diagnostic` and `recency`. The latter rule proceeded
without error if the metadata field was missing (albeit producing
a non-useful output) but the former required the field to exist. This
has prevented people from running the pipeline [1].

This commit allows the diagnostic step to proceed without the
'submission_date' field, printing a warning message instead.

[1] https://discussion.nextstrain.org/t/error-in-main-workflow-smk-file-in-latency-wait/616