docs: Note _why_ we request AWS/GCP compute loads to fetch from S3/GS
John (@huddlej) pointed out that this isn't obvious and providing a
reason may help this request "stick" in folks minds.

If it turns out that we're spending more than we want on transfer, we
can enable actual access controls to restrict access to the various
endpoints based on request source (e.g. within or outside of AWS).  For
now, let's just keep it simple by asking nicely.